ipcscale/sem-waitforzero: Add multiple threads/core
Add the ability to put multiple threads on the same core.
The new feature itself is not really useful, it's just
the first step for merging sem-waitforzero and sem-pingpong:

For a ping-pong test, it makes sense to either intentionally
put both partners on the same core (threadspercore=2) or to
put them on different cores (threadspercore=1).

Signed-off-by: Manfred Spraul <manfred@colorfullife.com>